refactor: NavHostViewModel stage 3 - extract conversation list state holder
